検出ロジック維持
