+++ packages/URLFinder/src/strategies/FallbackStrategy.ts
@@ -1,31 +1,17 @@
-/**
- * 最終フォールバック戦略
- * Step1、Step2が失敗した場合の最後の手段
- */
 
-import { ContactPageResult, SearchStrategy } from '../types/interfaces';
+import type { ContactPageResult, SearchStrategy } from '../types/interfaces';
+import { SearchState } from '../core/SearchState';
 
 export class FallbackStrategy implements SearchStrategy {
-
-  /**
-   * 戦略名を取得
-   */
-  getStrategyName(): string {
-    return 'Fallback Strategy (Final)';
+  public getStrategyName(): string {
+    return 'Fallback';
   }
 
-  /**
-   * 最終フォールバック検索の実行
-   */
-  search(_baseUrl: string): ContactPageResult | null {
-    console.log(`=== ${this.getStrategyName()} Starting ===`);
-    console.log('All primary strategies failed - returning not found result');
-    
-    return {
-      contactUrl: null,
-      actualFormUrl: null,
-      foundKeywords: ['not_found'],
-      searchMethod: 'not_found'
-    };
+  public search(baseUrl: string, searchState: SearchState): ContactPageResult | null {
+    const fallbackResult = searchState.getFinalResult();
+    if (fallbackResult.contactUrl) {
+      return fallbackResult;
+    }
+    return null;
   }
 }
